Moved the configuration for Bitcoin into pkg/config
