+++ pkg/config/config.go
@@ -4,8 +4,8 @@ import (
 	"github.com/keep-network/keep-core/pkg/bitcoin/electrum"
 )
 
-// Config defines the configuration for Bitcoin.
-type Config struct {
+// BitcoinConfig defines the configuration for Bitcoin.
+type BitcoinConfig struct {
 	// Electrum defines the configuration for the Electrum client.
 	Electrum electrum.Config
 }
